+++ src/examples/scala/com/decisionbrain/cplex/cp/SchedCumul.scala
@@ -50,7 +50,7 @@ object SchedCumul {
   def makeHouse(id: Int, releaseDate: Int): (IntExpr, IntervalVarArray) = {
 
     val taskVars: Map[String, IntervalVar] = (for (task <- tasks; (tname, tduration) = task)
-          yield (tname , model.intervalVar(sizeMin = tduration, sizeMax = tduration, name = "H" + id + "-" + tname)))(collection.breakOut)
+          yield (tname , model.intervalVar(sizeMin = tduration, sizeMax = tduration, name = "H" + id + "-" + tname))).toMap
 
     workersUsage += model.sum(for (e <- taskVars; (name, v) = e) yield pulse(v, 1))