+++ src/examples/scala/com/decisionbrain/cplex/mp/Diet.scala
@@ -74,7 +74,7 @@ object Diet {
 
     theFoodNutrients =
       (for (t <- foodNutients; n <- nutrients.indices)
-        yield (t._1, nutrients(n)._1) -> t.productElement(1 + n).asInstanceOf[Double])(collection.breakOut).toMap
+        yield (t._1, nutrients(n)._1) -> t.productElement(1 + n).asInstanceOf[Double]).toMap
 
 
     implicit val model = new MpModel("diet")