+++ src/test/scala/com/decisionbrain/cplex/mp/NursesTest.scala
@@ -24,6 +24,8 @@ class NursesTest extends FunSuite with Matchers {
     val model = nursesModel.buildModel()
     val status = nursesModel.solve(model)
 
+    implicit val order = Ordering.Double.IeeeOrdering
+
     model.getObjValue() should be <= 29975.5
     model.getValue(nursesModel.totalSalaryCost) should equal(29736.0 +- epsilon)
     model.getValue(nursesModel.totalNumberOfAssignments) should equal(224.0 +- epsilon)
@@ -38,6 +40,8 @@ class NursesTest extends FunSuite with Matchers {
 
     val model = nursesModel.buildModel()
     val status = nursesModel.solve(model)
+
+    implicit val order = Ordering.Double.IeeeOrdering
 
     model.getObjValue() should be <= 29975.5
     model.getValue(nursesModel.totalSalaryCost) should equal(29736.0 +- epsilon)